backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg[binary]==3.2.13
pgvector==0.2.4
sentence-transformers==3.0.1
beautifulsoup4==4.12.3
django-cors-headers==4.3.1
python-dotenv==1.0.0
